\name{IRISXin-package}
\alias{IRISXin-package}
\alias{IRISXin}
\docType{package}
\title{\packageTitle{IRISXin}}
\description{\packageDescription{IRISXin}}
\section{Package Content}{\packageIndices{IRISXin}}
\author{\packageAuthor{IRISXin}}
\section{Maintainer}{\packageMaintainer{IRISXin}}
\keyword{package}
